Text file and UI work
Fixed the name text fileds, and the time text filed by extending the text boxes to the side of the screen, and changing the text that gets shown.

Finalized the text file builder so that it makes the folder when the store opens, and saves a text file of all the clients for the day when the store closes.